pyproject.toml
[tool.poetry]
name = "fibsem-tools"
version = "6.3.2"
description = "Tools for processing FIBSEM datasets"
authors = ["Davis Vann Bennett <davis.v.bennett@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
distributed = ">=2021.10.0"
h5py = "^3.1.0"
fsspec = ">=2022.2.0"
zarr = "^2.10.3"
mrcfile = "^1.2.0"
numpy = "^1.20.0"
xarray = ">=2022.03.0"
pydantic = "^1.8.2"
backoff = "^1.10.0"
s3fs = ">=2022.2.0"
xarray-ome-ngff = "^1.2.1"
pint = "^0.20.1"
xarray-multiscale = "^2.0.0"
tifffile = "^2023.2.28"
pydantic-ome-ngff = "^0.3.0"
click = "^8.1.3"
dask = "^2023.3.2"
textual = "^0.16.0"
aiohttp = "^3.8.4"
xarray-datatree = "^0.0.12"
pydantic-zarr = "^0.5.0"
cellmap-schemas = ">=0.2.0"
pyairtable = "^2.1.0.post1"
python-dotenv = "^1.0.0"
imagecodecs = "^2024.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^3.0.0"
pre-commit = "2.21.0"
mypy = "^1.1.1"
requests = "^2.28.2"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.18"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}
pytest-examples = "^0.0.9"

[tool.poetry.scripts]
tiff2zarr = 'fibsem_tools.cli.tiff2zarr:run'
zarr-scan = 'fibsem_tools.cli.zarr_scan:cli'
zarr2json = 'fibsem_tools.cli.zarr2json:cli'
fst = 'fibsem_tools.cli.fst:fst'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "normal"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
